Try to fix deps & create IOSurface planes
My fault I'm still learning...

be nice in the comments 😘

+++ wgpu-core/src/device/global.rs
@@ -490,7 +490,7 @@ impl Global {
         (id, Some(error))
     }
 
-    pub fn texture_shared_handle(&self, texture_id: id::TextureId) -> Option<usize> {
+    pub fn texture_shared_handle(&self, texture_id: id::TextureId) -> Option<u32> {
         profiling::scope!("Texture::shared_handle");
         api_log!("Texture::shared_handle");
